Style changes, doc update

+++ src/gamma_main.c
@@ -6,11 +6,16 @@
 #include "gamma.h"
 #include "inputParser.h"
 
+#include "display.h"
+
+
 //Line buffer is dynamically allocated
 size_t INITIAL_BUFFER_SIZE = 0;
 
+//Globalne zmienne, by można było je zwolnić w exit_gamma
 gamma_t* g = NULL;
 char* line = NULL;
+
 
 void exit_gamma() {
     gamma_delete(g);
@@ -20,7 +25,7 @@ void exit_gamma() {
 int main() {
     atexit(exit_gamma);
 
-    unsigned long lineNum = 1;
+    unsigned long long lineNum = 1;
 
     while (getline(&line, &INITIAL_BUFFER_SIZE, stdin) > 0) {
         command_t command = getCommand(line);
